requirements.txt
requests
pymongo
pandas
matplotlib
seaborn
scikit-learn
python-dotenv

